moved the ui files into their own folder
