+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'home_page.dart';
+import 'ui/home_page.dart';
 
 void main() => runApp(const ChronoKeeper());
